requirements.txt
django~=3.2.21

djangorestframework~=3.13.1
markdown~=3.4.4
django-filter~=23.3
django-environ
django-celery-email
django-rest-swagger

celery[redis]~=5.3.4

psycopg2-binary
pip~=23.2.1
wheel~=0.41.2
amqp~=5.1.1
vine~=5.0.0
pytz~=2023.3.post1
kombu~=5.3.2
billiard~=4.1.0
redis~=4.6.0
requests~=2.31.0
setuptools~=68.2.2
click~=8.1.7
python-dateutil~=2.8.2
asgiref~=3.7.2
sqlparse~=0.4.4
colorama~=0.4.6
PyYAML~=6.0.1
wcwidth~=0.2.6
six~=1.16.0
packaging~=23.1
Cython~=3.0.2
platformdirs~=3.10.0
pyperclip~=1.8.2